[EI-243] [web] refactor: Refactor import paths for indicator-related services and view models
