+++ web/app/ui/components/numerical-guidance/indicator/advanced-indicators-chart.tsx
@@ -2,7 +2,7 @@ import { useHistoryIndicatorsValueViewModel } from '@/app/business/hooks/numeric
 import AdvancedMultiLineChart from '../../view/molecule/advanced-multi-line-chart/advanced-multi-line-chart';
 import { useLiveIndicatorsValueViewModel } from '@/app/business/hooks/numerical-guidance/indicator/use-live-indicators-value-view-model.hook';
 import { useEffect } from 'react';
-import { createIndicatorFormatter } from '@/app/business/services/chart/indicator-formatter.service';
+import { createIndicatorFormatter } from '@/app/business/services/numerical-guidance/chart/indicator-formatter.service';
 import { useCustomForecastIndicatorsValueViewModel } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicators-value-view-model.hook';
 import { useIndicatorBoardMetadataViewModel } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
