+++ web/app/ui/components/numerical-guidance/indicator/simple-indicators-chart.tsx
@@ -1,7 +1,7 @@
 import { useLiveIndicatorsValueViewModel } from '@/app/business/hooks/numerical-guidance/indicator/use-live-indicators-value-view-model.hook';
 import MultiLineChart from '../../view/molecule/multi-line-chart/multi-line-chart';
 import { useCustomForecastIndicatorsValueViewModel } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicators-value-view-model.hook';
-import { createIndicatorFormatter } from '@/app/business/services/chart/indicator-formatter.service';
+import { createIndicatorFormatter } from '@/app/business/services/numerical-guidance/chart/indicator-formatter.service';
 import { useIndicatorBoardMetadataViewModel } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
 
 type SimpleIndicatorsChartProps = {
